updt2.5D
i just deleted the player scene since we dont need it anymore and some temporary sprites that i used in that scene. Fixed the enemy collision but i still need to fix the button and making the enemy attack the player. I made the houses become a prehab so we can just reuse it.